[Feature] Provide more detailed debug print information in AcceptString and AcceptToken. (#450)
As reported in #443, it may be hard to understand the prediction and
completion relationship in the Earley parser. This PR provides more
detailed informations for developers and users to check the states in
the Earley parser.

This is the comparison between the main branch and this PR(left: this
PR, right: main branch).
<img width="2880" height="1620" alt="3979b53add5743a69d8337218aa6ca87"
src="https://github.com/user-attachments/assets/2d59e29e-eeee-4576-b1fc-883422b5b993"
/>

Signed-off-by: Yuchuan <yuchuan.7streams@gmail.com>